SNAP-370: HOI Performance
